feat(Serialization): Propose a first implementation of  ReferenceProperties.
First implementation of the reference properties.
A properties should not serialize its referenceProperties, but you will get the
type of the referenced properties instead.
The deserialization will not crash if a referenced properties is missing.
When you have deserialized all your object your can call the method
"JsonSchemaUtil.resolveReferenceProperties" to rebuild the complete object.
These method are going to be use to serialize and deserialize the Datasets
and the Datastores.